refactor: configure server to serve static files

+++ src/infrastructure/services/image-upload.service.ts
@@ -2,9 +2,17 @@ import { FileUplaodService } from '@modules/book/domain-services/file-upload.ser
 import * as fs from 'fs';
 import { imageStorageConfig } from '@config/storage.config';
 import { BookImage } from '@modules/book/domain/value-objects/image.value-object';
+import { ConflictException } from '@exceptions';
 
 export class ImageUploadService implements FileUplaodService {
   upload(file: Buffer, image: BookImage): void {
-    fs.writeFileSync(`${imageStorageConfig.dist}/${image.imageName}`, file);
+    // check if file exists
+    const filePath = `${__dirname}/../../upload/image/${image.imageName}`;
+    const exists = fs.existsSync(filePath);
+    if (exists) {
+      throw new ConflictException('File already exists');
+    } else {
+      fs.writeFileSync(`${imageStorageConfig.dist}/${image.imageName}`, file);
+    }
   }
 }
